Add working basic prompt generation
- Add cycling placeholder examples for prompt field
- Add 'Try Example' button to populate random example
- Fix Anthropic API type casting for max_tokens and temperature
- Add rainbow animated loading indicator during generation
- Working end-to-end HTML generation with Anthropic Claude API

+++ app/Services/AI/AnthropicClient.php
@@ -63,8 +63,8 @@ class AnthropicClient
                     ])
                     ->post('https://api.anthropic.com/v1/messages', [
                         'model' => config('mcp.anthropic.model'),
-                        'max_tokens' => config('mcp.anthropic.max_tokens', 4096),
-                        'temperature' => config('mcp.anthropic.temperature', 0.7),
+                        'max_tokens' => (int) config('mcp.anthropic.max_tokens', 4096),
+                        'temperature' => (float) config('mcp.anthropic.temperature', 0.7),
                         'system' => $systemPrompt,
                         'messages' => [
                             [
